Update dm_mripicture.py
add logging module and skip subjects with existing files

+++ bin/dm_mripicture.py
@@ -14,17 +14,27 @@ Options:
     -s --subject        Subjects
     -o --output=FOLDER  Output directory
     -t --tag=TAG        Scan tag [default: T1]
+    -r --replace        Replace existing files [default: False]
     -h --help           Show this screen
+    -q, --quiet         Show minimal output
+    -d, --debug         Show debug messages
+    -v, --verbose       Show intermediate steps
 """
 
 import os
 import glob
+import logging
 from nilearn import plotting
 import numpy as np
 from docopt import docopt
 
 import datman.config
 import datman.scan
+
+
+logging.basicConfig(level=logging.WARN,
+                    format="[%(name)s] %(levelname)s: %(message)s")
+logger = logging.getLogger(os.path.basename(__file__))
 
 
 def get_all_subjects(config):
@@ -40,20 +50,35 @@ def main():
     outdir = arguments['--output']
     subs = arguments['<subject>']
     tag = arguments['--tag']
+    overwrite = arguments['--replace']
+    quiet = arguments['--quiet']
+    debug = arguments['--debug']
+    verbose = arguments['--verbose']
     config = datman.config.config(study=study)
 
+    if verbose:
+        logger.setLevel(logging.INFO)
+    if debug:
+        logger.setLevel(logging.DEBUG)
+    if quiet:
+        logger.setLevel(logging.ERROR)
+
     if subs:
-        print('Creating brain pictures for subjects [', ', '.join(subs),
-              '] from', study, 'project using', tag, 'scans.')
+        logger.info('Creating pictures for subjects [ {} ] from {} project '
+                    'using {} scans.'.format(', '.join(subs), study, tag))
     else:
         subs = get_all_subjects(config)
-        print('Creating brain pictures for all subjects from',
-              study, 'project using', tag, 'scans.')
+        logger.info('Creating pictures for all {} subjects from {} project '
+                    'using {} scans.'.format(len(subs), study, tag))
 
     if not outdir:
         outdir = os.path.join(config.get_path('data'), 'tshirt')
 
     os.makedirs(outdir, exist_ok=True)
+    logger.debug('Output location set to: {}'.format(outdir))
+
+    if overwrite:
+        logger.info('Overwriting existing files')
 
     for subject in subs:
 
@@ -65,14 +90,20 @@ def main():
         imgpath = tagged_scan[idx].path
         outpath = os.path.join(outdir, subject + '_T1.pdf')
 
-        # Output Image
-        t1_pic = plotting.plot_anat(imgpath, cut_coords=(-20, -10, 2),
-                                    display_mode='x', annotate=False,
-                                    draw_cross=False, vmin=100,
-                                    vmax=1100, threshold='auto')
-        t1_pic.savefig(outpath, dpi=1000)
+        if os.path.isfile(outpath) and not overwrite:
+            logger.debug('Skipping subject {} as files already exist.'
+                         .format(subject))
+        else:
+            # Output Image
+            t1_pic = plotting.plot_anat(imgpath, cut_coords=(-20, -10, 2),
+                                        display_mode='x', annotate=False,
+                                        draw_cross=False, vmin=100,
+                                        vmax=1100, threshold='auto')
+            t1_pic.savefig(outpath, dpi=1000)
+            logger.debug('Created new brain pictures for subject {} from file '
+                         '{} and saved as {}'.format(subject, imgpath, outpath))
 
-    print('Saved all output to', outdir)
+    logger.info('Saved all output to: {}'.format(outdir))
 
 
 if __name__ == "__main__":
